requirements.txt
slack-bolt
feedparser
requests
apscheduler
python-dotenv
openai
google-generativeai
